Create README.md
Loan management is an essential aspect of banking that plays a crucial role in ensuring the stability and profitability of financial institutions A loan management system enables lenders to seamlessly manage the loan lifecycle, centralizes data, enhances communication, and accelerates decision-making, significantly improving operational efficiency. This case study aims to give us an idea of applying Expository Data Analysis in a real business scenario and understand how data is used to minimize the risk of losing money while lending to customers.